requirements.txt
streamlit
openai
pytesseract
Pillow
scikit-learn
requests
beautifulsoup4
python-dotenv
